Create CLIP API
